fix(ocp): utilisation du double systématiquement

+++ src/solid/OCP/AreaCalculator.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public class AreaCalculator {
 
-	public float area(List<Rectangle> rectangles) {
-		float area = 0;
+	public double area(List<Rectangle> rectangles) {
+		double area = 0;
 		
 		for (Rectangle rectangle : rectangles) {
 			area += rectangle.getHeight() * rectangle.getWidth();
